.github/scripts/requirements.txt
google-auth-oauthlib>=0.4.6
google-auth>=2.0.0
google-api-python-client>=2.0.0